Модуль utils.js разбит на два новых модуля: dom и random
